+++ test/size/vue/responsive/main.ts
@@ -1,3 +1,3 @@
-import { responsive } from '@nuxt-photo/vue'
+import { responsive } from '@lupinum/vue-photo'
 
 console.log(responsive({ 0: 2, 768: 3 }))